+++ cli/cli/command_framework/highlevel/enclave_id_arg/enclave_id_arg.go
@@ -2,8 +2,8 @@ package enclave_id_arg
 
 import (
 	"context"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command/args"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command/flags"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/lowlevel/args"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/lowlevel/flags"
 	"github.com/kurtosis-tech/kurtosis-engine-api-lib/api/golang/kurtosis_engine_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-engine-api-lib/api/golang/lib/kurtosis_context"
 	"github.com/kurtosis-tech/stacktrace"