+++ cli/cli/command_framework/lowlevel/must_get_cobra_command_test.go
@@ -1,8 +1,8 @@
-package kurtosis_command
+package lowlevel
 
 import (
-	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command/args"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command/flags"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/lowlevel/args"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/lowlevel/flags"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
@@ -43,16 +43,16 @@ var validTokens = []string{
 }
 
 func TestMustGetCobraCommand_DuplicateArgsCausePanic(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
 		Args:             []*args.ArgConfig{
 			{
-				Key:             arg1Key,
+				Key: arg1Key,
 			},
 			{
-				Key:             arg1Key,
+				Key: arg1Key,
 			},
 		},
 	}
@@ -65,7 +65,7 @@ func TestMustGetCobraCommand_DuplicateArgsCausePanic(t *testing.T) {
 }
 
 func TestMustGetCobraCommand_DuplicateFlagsCausePanic(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
@@ -100,7 +100,7 @@ func TestMustGetCobraCommand_FlagsWithMismatchedDefaulValuesCausePanic(t *testin
 		},
 	}
 	for _, illegalFlag := range illegalFlagVariants {
-		kurtosisCmd := &KurtosisCommand{
+		kurtosisCmd := &LowlevelKurtosisCommand{
 			CommandStr:       "test",
 			ShortDescription: "Short description",
 			LongDescription:  "This is a very long description",
@@ -120,13 +120,13 @@ func TestMustGetCobraCommand_FlagsWithMismatchedDefaulValuesCausePanic(t *testin
 }
 
 func TestMustGetCobraCommand_FlagWithTypeDoesntPanic(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
 		Flags: []*flags.FlagConfig{
 			{
-				Key: flag1Key,
+				Key:  flag1Key,
 				Type: flags.FlagType_String,
 			},
 		},
@@ -141,17 +141,17 @@ func TestMustGetCobraCommand_FlagWithTypeDoesntPanic(t *testing.T) {
 
 func TestMustGetCobraCommand_DuplicateFlagShorthandsPanic(t *testing.T) {
 	dupedShorthandValue := "x"
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
 		Flags: []*flags.FlagConfig{
 			{
-				Key: flag1Key,
+				Key:       flag1Key,
 				Shorthand: dupedShorthandValue,
 			},
 			{
-				Key: flag2Key,
+				Key:       flag2Key,
 				Shorthand: dupedShorthandValue,
 			},
 		},
@@ -165,13 +165,13 @@ func TestMustGetCobraCommand_DuplicateFlagShorthandsPanic(t *testing.T) {
 }
 
 func TestMustGetCobraCommand_ShorthandsGreaterThanOneLetterPanic(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
 		Flags: []*flags.FlagConfig{
 			{
-				Key: flag1Key,
+				Key:       flag1Key,
 				Shorthand: "this is way too long",
 			},
 		},
@@ -185,17 +185,17 @@ func TestMustGetCobraCommand_ShorthandsGreaterThanOneLetterPanic(t *testing.T) {
 }
 
 func TestMustGetCobraCommand_TestEmptyShorthandsDontTriggerShorthandValidation(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
 		Flags: []*flags.FlagConfig{
 			{
-				Key: flag1Key,
+				Key:       flag1Key,
 				Shorthand: "",
 			},
 			{
-				Key: flag2Key,
+				Key:       flag2Key,
 				Shorthand: "",
 			},
 		},
@@ -209,7 +209,7 @@ func TestMustGetCobraCommand_TestEmptyShorthandsDontTriggerShorthandValidation(t
 }
 
 func TestMustGetCobraCommand_EmptyArgKeyCausesPanic(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
@@ -228,7 +228,7 @@ func TestMustGetCobraCommand_EmptyArgKeyCausesPanic(t *testing.T) {
 }
 
 func TestMustGetCobraCommand_EmptyFlagKeyCausesPanic(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
@@ -247,20 +247,20 @@ func TestMustGetCobraCommand_EmptyFlagKeyCausesPanic(t *testing.T) {
 }
 
 func TestMustGetCobraCommand_TwoOptionalArgumentsCausePanic(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
 		Args:             []*args.ArgConfig{
 			{
-				Key:             arg1Key,
+				Key: arg1Key,
 			},
 			{
-				Key:             arg2Key,
+				Key:        arg2Key,
 				IsOptional: true,
 			},
 			{
-				Key:             arg3Key,
+				Key:        arg3Key,
 				IsOptional: true,
 			},
 		},
@@ -274,20 +274,20 @@ func TestMustGetCobraCommand_TwoOptionalArgumentsCausePanic(t *testing.T) {
 }
 
 func TestMustGetCobraCommand_MiddleGreedyArgCausesPanic(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
 		Args:             []*args.ArgConfig{
 			{
-				Key:             arg1Key,
+				Key: arg1Key,
 			},
 			{
-				Key:             arg2Key,
+				Key:      arg2Key,
 				IsGreedy: true,
 			},
 			{
-				Key:             arg3Key,
+				Key: arg3Key,
 			},
 		},
 	}
@@ -300,7 +300,7 @@ func TestMustGetCobraCommand_MiddleGreedyArgCausesPanic(t *testing.T) {
 }
 
 func TestMustGetCobraCommand_WorkingFlagDefaultValueChecking(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
@@ -326,15 +326,15 @@ func TestMustGetCobraCommand_WorkingFlagDefaultValueChecking(t *testing.T) {
 }
 
 func TestMustGetCobraCommand_OptionalArgsWithNilDefaultPanic(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
 		Args: []*args.ArgConfig{
 			{
-				Key: arg1Key,
+				Key:          arg1Key,
 				DefaultValue: nil,
-				IsOptional: true,
+				IsOptional:   true,
 			},
 		},
 	}
@@ -347,15 +347,15 @@ func TestMustGetCobraCommand_OptionalArgsWithNilDefaultPanic(t *testing.T) {
 }
 
 func TestMustGetCobraCommand_RequiredArgWithNilDefaultDoesntPanic(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
 		Args: []*args.ArgConfig{
 			{
-				Key: arg1Key,
+				Key:          arg1Key,
 				DefaultValue: nil,
-				IsOptional: false,
+				IsOptional:   false,
 			},
 		},
 	}
@@ -368,16 +368,16 @@ func TestMustGetCobraCommand_RequiredArgWithNilDefaultDoesntPanic(t *testing.T) 
 }
 
 func TestMustGetCobraCommand_OptionalNonGreedyArgWithWrongDefaultTypePanics(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
 		Args: []*args.ArgConfig{
 			{
-				Key: arg1Key,
-				IsGreedy: false,
+				Key:          arg1Key,
+				IsGreedy:     false,
 				DefaultValue: []string{"foo", "bar"},
-				IsOptional: true,
+				IsOptional:   true,
 			},
 		},
 	}
@@ -390,16 +390,16 @@ func TestMustGetCobraCommand_OptionalNonGreedyArgWithWrongDefaultTypePanics(t *t
 }
 
 func TestMustGetCobraCommand_OptionalGreedyArgWithWrongDefaultTypePanics(t *testing.T) {
-	kurtosisCmd := &KurtosisCommand{
+	kurtosisCmd := &LowlevelKurtosisCommand{
 		CommandStr:       "test",
 		ShortDescription: "Short description",
 		LongDescription:  "This is a very long description",
 		Args: []*args.ArgConfig{
 			{
-				Key: arg1Key,
-				IsGreedy: true,
+				Key:          arg1Key,
+				IsGreedy:     true,
 				DefaultValue: "foobar",
-				IsOptional: true,
+				IsOptional:   true,
 			},
 		},
 	}