+++ cli/cli/commands/enclave/rm/rm.go
@@ -4,17 +4,13 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/docker/docker/client"
 	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command/args"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command/flags"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/prebuilt_command_components/enclave_id_arg"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/highlevel/enclave_id_arg"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/highlevel/engine_consuming_kurtosis_command"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/lowlevel/args"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/lowlevel/flags"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/defaults"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/engine_manager"
 	"github.com/kurtosis-tech/kurtosis-engine-api-lib/api/golang/kurtosis_engine_rpc_api_bindings"
-	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"google.golang.org/protobuf/types/known/emptypb"
@@ -23,23 +19,24 @@ import (
 )
 
 const (
-	enclaveIdArgKey          = "enclave-id"
-	shouldForceRemoveFlagKey = "force"
-
-	engineClientCtxKey = "engine-client"
-	engineClientCloseFuncCtxKey = "engine-client-close-func"
-
+	enclaveIdArgKey = "enclave-id"
 	isEnclaveIdArgOptional = false
 	isEnclaveIdArgGreedy = true
 
+	shouldForceRemoveFlagKey = "force"
 	defaultShouldForceRemove = "false"
+
+	dockerManagerCtxKey = "docker-manager"
+	engineClientCtxKey = "engine-client"
 )
 
-var EnclaveRmCmd = &kurtosis_command.KurtosisCommand{
-	CommandStr:       command_str_consts.EnclaveRmCmdStr,
-	ShortDescription: "Destroys the specified enclaves",
-	LongDescription:  "Destroys the specified enclaves, removing all resources associated with them",
-	Flags:            []*flags.FlagConfig{
+var EnclaveRmCmd = &engine_consuming_kurtosis_command.EngineConsumingKurtosisCommand{
+	CommandStr:              command_str_consts.EnclaveRmCmdStr,
+	ShortDescription:        "Destroys the specified enclaves",
+	LongDescription:         "Destroys the specified enclaves, removing all resources associated with them",
+	DockerManagerContextKey: dockerManagerCtxKey,
+	EngineClientContextKey:  engineClientCtxKey,
+	Flags: []*flags.FlagConfig{
 		{
 			Key:       shouldForceRemoveFlagKey,
 			Usage:     "Deletes all enclaves, regardless of whether they're already stopped",
@@ -48,7 +45,7 @@ var EnclaveRmCmd = &kurtosis_command.KurtosisCommand{
 			Default:   defaultShouldForceRemove,
 		},
 	},
-	Args:             []*args.ArgConfig{
+	Args:                    []*args.ArgConfig{
 		enclave_id_arg.NewEnclaveIDArg(
 			enclaveIdArgKey,
 			engineClientCtxKey,
@@ -56,59 +53,30 @@ var EnclaveRmCmd = &kurtosis_command.KurtosisCommand{
 			isEnclaveIdArgGreedy,
 		),
 	},
-	PreValidationAndRunFunc: setup,
-	RunFunc:          run,
-	PostValidationAndRunFunc: teardown,
+	RunFunc: run,
 }
 
-// TODO generalize this to something like EngineUsingKurtosisCommand or something
-func setup(ctx context.Context) (context.Context, error) {
-	result := ctx
-
-	dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
+func run(
+	ctx context.Context,
+	dockerManager *docker_manager.DockerManager,
+	engineClient kurtosis_engine_rpc_api_bindings.EngineServiceClient,
+	flags *flags.ParsedFlags,
+	args *args.ParsedArgs,
+) error {
+	enclaveIds, err := args.GetGreedyArg(enclaveIdArgKey)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred creating the Docker client")
-	}
-	dockerManager := docker_manager.NewDockerManager(
-		logrus.StandardLogger(),
-		dockerClient,
-	)
-
-	engineManager := engine_manager.NewEngineManager(dockerManager)
-	objAttrsProvider := schema.GetObjectAttributesProvider()
-	engineClient, closeClientFunc, err := engineManager.StartEngineIdempotentlyWithDefaultVersion(ctx, objAttrsProvider, defaults.DefaultEngineLogLevel)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred creating a new Kurtosis engine client")
-	}
-	result = context.WithValue(result, engineClientCtxKey, engineClient)
-	result = context.WithValue(result, engineClientCloseFuncCtxKey, closeClientFunc)
-
-	return result, nil
-}
-
-func run(ctx context.Context, flags *flags.ParsedFlags, args *args.ParsedArgs) error {
-	uncastedEngineClient := ctx.Value(engineClientCtxKey)
-	if uncastedEngineClient == nil {
-		return stacktrace.NewError("Expected an engine client to have been stored in the context under key '%v', but none was found; this is a bug in Kurtosis!", engineClientCtxKey)
-	}
-	engineClient, ok := uncastedEngineClient.(kurtosis_engine_rpc_api_bindings.EngineServiceClient)
-	if !ok {
-		return stacktrace.NewError("Found an object that should be the engine client stored in the context under key '%v', but this object wasn't of the correct type", engineClientCtxKey)
+		return stacktrace.Propagate(err, "Expected a value for greedy enclave ID arg '%v' but none was found; this is a bug with Kurtosis!", enclaveIdArgKey)
 	}
 
-	inputtedEnclaveIds, err := args.GetGreedyArg(enclaveIdArgKey)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting the enclave IDs using arg key '%v'; this is a bug in Kurtosis!", enclaveIdArgKey)
-	}
 	shouldForceRemove, err := flags.GetBool(shouldForceRemoveFlagKey)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred getting the force-removal flag value using key '%v'; this is a bug in Kurtosis!", shouldForceRemoveFlagKey)
 	}
 
-	logrus.Debugf("inputted enclave IDs: %+v", inputtedEnclaveIds)
+	logrus.Debugf("inputted enclave IDs: %+v", enclaveIds)
 
 	// Condense the enclave IDs down into a unique set, so we don't try to double-destroy an enclave
-	enclaveIdsToDestroy := getUniqueSortedEnclaveIDs(inputtedEnclaveIds)
+	enclaveIdsToDestroy := getUniqueSortedEnclaveIDs(enclaveIds)
 
 	logrus.Debugf("Unique enclave IDs to destroy: %+v", enclaveIdsToDestroy)
 
@@ -144,25 +112,6 @@ func run(ctx context.Context, flags *flags.ParsedFlags, args *args.ParsedArgs) e
 	logrus.Info("Enclaves successfully destroyed")
 
 	return nil
-}
-
-func teardown(ctx context.Context) {
-	uncastedEngineClientCloseFunc := ctx.Value(engineClientCloseFuncCtxKey)
-	if uncastedEngineClientCloseFunc != nil {
-		engineClientCloseFunc, ok := uncastedEngineClientCloseFunc.(func() error)
-		if ok {
-			if err := engineClientCloseFunc(); err != nil {
-				logrus.Warnf("We tried to close the engine client after we're done using it, but doing so threw an error:\n%v", err)
-			}
-		} else {
-			logrus.Errorf("Expected the object at context key '%v' to be an engine client close function, but it wasn't; this is a bug in Kurtosis!", engineClientCloseFuncCtxKey)
-		}
-	} else {
-		logrus.Errorf(
-			"Expected to find an engine client close function during teardown at context key '%v', but none was found; this is a bug in Kurtosis!",
-			engineClientCloseFuncCtxKey,
-		)
-	}
 }
 
 // ====================================================================================================